refactor: move rules from phi-paper
