+++ eo-phi-normalizer/src/Language/EO/Phi/Rules/Common.hs
@@ -78,20 +78,6 @@ applyRulesChain ctx obj
       | obj' <- applyOneRule ctx obj
       , chain <- applyRulesChain ctx obj' ]
 
--- * Yegor's Rules
-
--- | Rule 6.
-rule6 :: Rule
-rule6 ctx (ObjectDispatch (Formation bindings) a)
-  | Just obj <- lookupBinding a bindings = do
-    guard (isNF ctx obj)
-    return (Application obj [AlphaBinding Rho (Formation bindings')])
-    where
-      bindings' = filter (not . isDispatched) bindings
-      isDispatched (AlphaBinding a' _) = a == a'
-      isDispatched _ = False
-rule6 _ _ = Nothing
-
 -- * Helpers
 
 -- | Lookup a binding by the attribute name.
